Support component matches by code element for `UsesComponent` (Cecil)
Try to find the destination component for a relationship by the main
type of the component, but if we can't locate such a component then
try looking through the code elements for each component in the
container. If there are multiple matches via code element, abandon the
match attempt and warn the user that there were multiple possibilities.